Break out GraphQLCompiler and TemplateString
The “Frontend” is now contained in it’s own target named `GraphQLCompiler` and TemplateString is also a new target. These targets are part of the `ApolloCodgenLib` package. But they are not standalone products and cannot be consumed externally.

This also changes the `auto_rollup` script used for compiling the JS Frontend. It changes the paths of the distribution to be inside of the “JavaScript” folder so it can more easily be excluded from the package files.