Updating docs publish job
Currently the docs publish job triggers twice for every update because its based on pushes to the main branch and our workflow updates and pushes the commit message when PRs are merged triggering a second run. Updating this to be based on merged PRs so that it will only trigger once like our subtree job.