Make required mock fields nonoptional in convenience initializer
Makes the mock initialization parameters corresponding to nonull
fields nonoptional. This allows the compiler to detect an invalid
mock, rather than allowing an incomplete mock to cause an error
at runtime.

Also fixes a bug in which the rendered type of a deeply nested
mock field would be typed as required, even though it might be
defined as optional.