Make cache public within ReadTransaction
Some users have use cases for accessing a custom `NormalizedCache` implementation directly while performing cache transactions. We don’t want to make `cache` a `public` property on the `ApolloStore`, because that would allow users to manipulate the cache outside of the Reader/Writer lock (implemented by the barrier queue). Making this `public` on the `ReadTransaction` (and therby it’s subclass `ReadWriteTransaction`), we can allow custom cache access only when it is safe to do so from within a transaction.

By adding a new `ReadOnlyNormalizedCache` protocol, we can expose the cache as read-only in the `ReadTransaction` and as writable in the `ReadWriteTransaction`.