Add subscript for optional values to InputObject
Optional values are only used for input objects with a schema defined default value. Otherwise they are GraphQLNullable. In 2.0 we remove conformance of `Optional` to `GraphQLOperationVariableValue` because it should always be stored as a `GraphQLNullable`.

This adds a new subscript to `InputDict` that converts the `GraphQLNullable` to the `Optional` field value (and vice versa for the setter).