Fix SelectionSet Equality bug with child merged field from fragment
This fixes https://github.com/apollographql/apollo-ios/issues/3602

When a field merged from a child object in a named fragment was merged, it was not being used for equality checks because it is not included in the direct selections of the object. The parent where the fragment is selected would have these selections, but the child object doesn’t. The only place that indicates these merged fields exist is the `__fulfilledFragments`.

Equality checks now use the types `__fulfilledFragments` and collects selections from all of them, instead of just traversing downward from `self.__selections`. This ensures all expected data is collected from merged fragments.

This PR also adds a bit of memoization by keeping track of which fragments have already been added to prevent unneeded duplicate traversal of the same fragment.